+++ src/infill.cpp
@@ -37,7 +37,7 @@ void generateConcentricInfill(Polygons outline, Polygons& result, int inset_valu
 
 
 void generateGridInfill(const Polygons& in_outline, int outlineOffset, Polygons& result,
-                        int extrusionWidth, int lineSpacing, int infillOverlap,
+                        int extrusionWidth, int lineSpacing, double infillOverlap,
                         double rotation)
 {
     generateLineInfill(in_outline, outlineOffset, result, extrusionWidth, lineSpacing,
@@ -47,7 +47,7 @@ void generateGridInfill(const Polygons& in_outline, int outlineOffset, Polygons&
 }
 
 void generateTriangleInfill(const Polygons& in_outline, int outlineOffset, Polygons& result,
-                        int extrusionWidth, int lineSpacing, int infillOverlap,
+                        int extrusionWidth, int lineSpacing, double infillOverlap,
                         double rotation)
 {
     generateLineInfill(in_outline, outlineOffset, result, extrusionWidth, lineSpacing,
@@ -107,7 +107,7 @@ void addLineInfill(Polygons& result, PointMatrix matrix, int scanline_min_idx, i
  *      and connect them using the even-odd rule
  * 
  */
-void generateLineInfill(const Polygons& in_outline, int outlineOffset, Polygons& result, int extrusionWidth, int lineSpacing, int infillOverlap, double rotation)
+void generateLineInfill(const Polygons& in_outline, int outlineOffset, Polygons& result, int extrusionWidth, int lineSpacing, double infillOverlap, double rotation)
 {
     if (in_outline.size() == 0) return;
     Polygons outline = in_outline.offset(extrusionWidth * infillOverlap / 100 + outlineOffset);
@@ -164,7 +164,7 @@ void generateLineInfill(const Polygons& in_outline, int outlineOffset, Polygons&
 }
 
 
-void generateZigZagInfill(const Polygons& in_outline, Polygons& result, int extrusionWidth, int lineSpacing, int infillOverlap, double rotation, bool connect_zigzags, bool use_endPieces)
+void generateZigZagInfill(const Polygons& in_outline, Polygons& result, int extrusionWidth, int lineSpacing, double infillOverlap, double rotation, bool connect_zigzags, bool use_endPieces)
 {
     if (use_endPieces) return generateZigZagInfill_endPieces(in_outline, result, extrusionWidth, lineSpacing, infillOverlap, rotation, connect_zigzags);
     else return generateZigZagInfill_noEndPieces(in_outline, result, extrusionWidth, lineSpacing, infillOverlap, rotation);
@@ -219,7 +219,7 @@ void generateZigZagInfill(const Polygons& in_outline, Polygons& result, int extr
  *   ^     ^     ^    scanlines
  *                 ^  disconnected end piece
  */
-void generateZigZagInfill_endPieces(const Polygons& in_outline, Polygons& result, int extrusionWidth, int lineSpacing, int infillOverlap, double rotation, bool connect_zigzags)
+void generateZigZagInfill_endPieces(const Polygons& in_outline, Polygons& result, int extrusionWidth, int lineSpacing, double infillOverlap, double rotation, bool connect_zigzags)
 {
 //     if (in_outline.size() == 0) return;
 //     Polygons outline = in_outline.offset(extrusionWidth * infillOverlap / 100 - extrusionWidth / 2);
@@ -353,7 +353,7 @@ void generateZigZagInfill_endPieces(const Polygons& in_outline, Polygons& result
 }
 
 
-void generateZigZagInfill_noEndPieces(const Polygons& in_outline, Polygons& result, int extrusionWidth, int lineSpacing, int infillOverlap, double rotation)
+void generateZigZagInfill_noEndPieces(const Polygons& in_outline, Polygons& result, int extrusionWidth, int lineSpacing, double infillOverlap, double rotation)
 {
     if (in_outline.size() == 0) return;
     Polygons outline = in_outline.offset(extrusionWidth * infillOverlap / 100 - extrusionWidth / 2);
